requirements.txt
asgiref==3.4.1
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
decorator==5.0.9
Django==3.2.6
django-phonenumber-field==5.2.0
django-sendsms==0.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
humanize==3.11.0
idna==3.2
ipython==7.27.0
jedi==0.18.0
matplotlib-inline==0.1.2
mysqlclient==2.0.3
parso==0.8.2
pexpect==4.8.0
phonenumbers==8.12.31
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.1
traitlets==5.1.0
urllib3==1.26.6
wcwidth==0.2.5
yapf==0.31.0
